Phase 6d: carve the results domain out of app.js into results.ts
results.ts owns the final-results screen (tables + champion crown,
battle graphic, round-history accordion with judge votes and Spotify
enrichment), the router hydration for /results (showResults +
hydrateResultsRoute via the typed getResults client), the generalized
battle-payload editor modal, the social-image canvas export, and the
battle JSON download. It also owns the state: currentResultsData, the
uploaded hustlentussle.battle payload, editor payload/callbacks, and
the social-export caches (the never-read resultsNumRounds cache was
dropped).

app.js supplies live-game accessors via initResultsDeps (sessionId,
guestJudges, and a scoreboard getter/setter — currentLeads/Follows are
shared with the live battle screen) and calls initResultsUI() for the
DOM wiring, which moved here wholesale (edit modal buttons + live
input/change consistency-warning refresh, download, social export).
window.openBattlePayloadEditor stays exported for ytd.ts; its declared
onSave meta is now correctly nullable (null when showMetaFields is
false), with a guard added at ytd.ts's call site. Champions.lead/follow
became nullable in types.ts to match web/helpers.py.

Verified: build + tsc + eslint clean; 9-step Playwright smoke passed;
targeted browser check covering live results (crown/graphic/accordion,
edit hidden), JSON download, social-image export, re-upload (edit
visible), editor rename + save + re-render, and the live points-drift
consistency warning — zero JS errors. window.openBattlePayloadEditor
confirmed present post-bundle.
