+++ web/js/app.js
@@ -15,6 +15,27 @@ import {
 } from './spotify';
 import { initDemo, currentDemoAction, demoActionCompleted } from './demo';
 import { initTiebreakDeps, startTiebreakFlow } from './tiebreak';
+import {
+    initResultsDeps,
+    initResultsUI,
+    showResults,
+    hydrateResultsRoute,
+    displayResults,
+    setUploadedBattlePayload,
+    resetResultsState,
+} from './results';
+
+// Wire the results domain to the live-game state it still reads/writes
+// (function declarations and `let` bindings hoist; arrows defer evaluation)
+initResultsDeps({
+    getSessionId: () => sessionId,
+    getGuestJudges: () => guestJudges,
+    setScoreboard: (leads, follows) => {
+        currentLeads = leads;
+        currentFollows = follows;
+    },
+    getScoreboard: () => ({ leads: currentLeads, follows: currentFollows }),
+});
 
 // Wire the tie-break modal to app state (function declarations hoist)
 initTiebreakDeps({
@@ -54,16 +75,7 @@ initDisplayDeps({
 // Global variables
 let sessionId = null;
 let liveBattleActive = false;      // a battle is live in-memory (skip refetch on route enter)
-let currentResultsData = null;     // last rendered results payload (for /results hydration)
-let currentUploadedBattlePayload = null; // raw hustlentussle.battle payload for the uploaded file being viewed/edited
-
-// State for the generalized battle-payload editor (#edit-results-modal). Kept separate from
-// currentUploadedBattlePayload since the same modal is also used to edit already-published
-// YTD battles (see web/js/ytd.js) - reusing that variable here would corrupt whichever flow
-// isn't currently using the modal (the results-screen download button, the edit-button
-// visibility gate, etc. all key off currentUploadedBattlePayload).
-let editorPayload = null;
-let editorCallbacks = null;
+// Results state (rendered payload, uploaded battle, editor modal) lives in results.ts
 let guestJudges = [];
 let leadVotes = {};  // Changed to an object to easily update votes
 let followVotes = {}; // Changed to an object to easily update votes
@@ -74,17 +86,6 @@ let votingLocked = { lead: false, follow: false }; // Track if voting is locked
 let initialLeads = []; // Store initial order of leads
 let initialFollows = []; // Store initial order of follows
 let contestantJudgingEnabled = true; // Track whether contestant judging is enabled for the battle
-
-// Cached data for the social export image (populated by displayResults)
-let resultsLeadMap = new Map();
-let resultsFollowMap = new Map();
-let resultsInitialLeads = [];
-let resultsInitialFollows = [];
-let resultsTopLeadName = null;
-let resultsTopFollowName = null;
-let resultsNumRounds = 0;
-let resultsGuestJudges = [];
-let resultsEventTitle = null; // custom subtitle; falls back to "Month Edition (Year)"
 
 // Display mode state (for viewer-only mode without voting controls)
 // Display mode state lives in display.ts (`display.active` + polling internals)
@@ -167,7 +168,7 @@ let submitVotesBtn, votingResults;
 let leadWinnerPreview, followWinnerPreview, leadPreviewName, followPreviewName;
 let roundResultsSection, winMessages, nextRoundBtn, endBattleBtn;
 let leadsLeaderboard, followsLeaderboard;
-let backToHomeFromResultsBtn, downloadBattleDataBtn, exportSocialImageBtn;
+let backToHomeFromResultsBtn;
 // Vote confirmation modal elements
 let voteConfirmModal, voteConfirmCloseBtn, voteConfirmCancelBtn, voteConfirmSubmitBtn;
 let voteConfirmRound, voteConfirmLead1, voteConfirmLead2, voteConfirmFollow1, voteConfirmFollow2;
@@ -350,8 +351,6 @@ document.addEventListener('DOMContentLoaded', () => {
     leadsLeaderboard = document.getElementById('leads-leaderboard');
     followsLeaderboard = document.getElementById('follows-leaderboard');
     backToHomeFromResultsBtn = document.getElementById('back-to-home-from-results');
-    downloadBattleDataBtn = document.getElementById('download-battle-data');
-    exportSocialImageBtn = document.getElementById('export-social-image');
     
     // Check if elements exist
     console.log('Checking elements:');
@@ -466,55 +465,8 @@ document.addEventListener('DOMContentLoaded', () => {
         console.error('backToHomeFromResultsBtn is null or undefined!');
     }
     
-    downloadBattleDataBtn.addEventListener('click', downloadBattleData);
-    if (exportSocialImageBtn) exportSocialImageBtn.addEventListener('click', exportSocialImage);
-
-    // Edit Results modal (uploaded battles only)
-    const editResultsBtn = document.getElementById('edit-results-btn');
-    const editResultsModal = document.getElementById('edit-results-modal');
-    const editResultsCloseBtn = document.getElementById('edit-results-close');
-    const editResultsCancelBtn = document.getElementById('edit-results-cancel');
-    const editResultsSaveBtn = document.getElementById('edit-results-save');
-    const editAddLeadBtn = document.getElementById('edit-add-lead');
-    const editAddFollowBtn = document.getElementById('edit-add-follow');
-    if (editResultsBtn) {
-        editResultsBtn.addEventListener('click', () => {
-            openEditResultsModal(currentUploadedBattlePayload, {
-                title: 'Edit Results',
-                showMetaFields: false,
-                onSave: async (payload) => {
-                    const reprocessed = await reprocessBattlePayload(payload);
-                    currentUploadedBattlePayload = payload;
-                    reprocessed.uploaded = true;
-                    showResults(reprocessed, { sessionId: null });
-                    showToast('Results updated', 'success');
-                },
-            });
-        });
-    }
-    if (editResultsCloseBtn) editResultsCloseBtn.addEventListener('click', closeEditResultsModal);
-    if (editResultsCancelBtn) editResultsCancelBtn.addEventListener('click', closeEditResultsModal);
-    if (editResultsSaveBtn) editResultsSaveBtn.addEventListener('click', saveEditedResults);
-    if (editAddLeadBtn) {
-        editAddLeadBtn.addEventListener('click', () => {
-            document.getElementById('edit-leads-body').appendChild(buildEditParticipantRow({ name: '', points: 0 }));
-            refreshEditConsistencyWarnings();
-        });
-    }
-    if (editAddFollowBtn) {
-        editAddFollowBtn.addEventListener('click', () => {
-            document.getElementById('edit-follows-body').appendChild(buildEditParticipantRow({ name: '', points: 0 }));
-            refreshEditConsistencyWarnings();
-        });
-    }
-    if (editResultsModal) {
-        editResultsModal.addEventListener('click', (e) => {
-            if (e.target === editResultsModal) closeEditResultsModal();
-        });
-        // Live-refresh the points-vs-round-wins warning as points/round winners are edited
-        editResultsModal.addEventListener('input', refreshEditConsistencyWarnings);
-        editResultsModal.addEventListener('change', refreshEditConsistencyWarnings);
-    }
+    // Results screen + edit-results modal wiring lives in results.ts
+    initResultsUI();
 
     // Theme toggles (nav bar + floating for display mode)
     document.querySelectorAll('#theme-toggle, #theme-toggle-floating').forEach(btn => {
@@ -1254,7 +1206,7 @@ async function processUploadedFile() {
         }
         
         // Display the results
-        currentUploadedBattlePayload = rawPayload;
+        setUploadedBattlePayload(rawPayload);
         data.uploaded = true;
         showResults(data, { sessionId: null });
     } catch (error) {
@@ -2202,8 +2154,7 @@ function resetAndGoHome() {
     console.log('resetAndGoHome called');
     resetCompetition();
     liveBattleActive = false;
-    currentResultsData = null;
-    currentUploadedBattlePayload = null;
+    resetResultsState();
     console.log('resetCompetition completed, navigating home');
     navigate('/');
 }
@@ -2290,400 +2241,7 @@ function updateScoreTable(leads, follows) {
     });
 }
 
-// --- Edit Results (uploaded battles only) ---------------------------------
-//
-// Uploaded battles never touch the server-side repository (see
-// /api/process_uploaded_file), so "editing" means mutating the raw
-// hustlentussle.battle payload we kept in memory and re-running it through
-// the same backend endpoint used for the original upload — that guarantees
-// the display shape (medals, placements) stays in sync with the same rules
-// a fresh upload would get.
-
-// Each round win is worth exactly 1 point (see game_logic.py: `winner.points += 1`,
-// ties/no-contests award none), so a contestant's recorded points should always equal
-// how many rounds they won. Uploaded/edited data can drift from that; flag it rather
-// than silently trusting whichever field was typed in.
-function computeRoundWinCounts(rounds, winnerKey) {
-    const wins = {};
-    (rounds || []).forEach((r) => {
-        const winner = r ? r[winnerKey] : null;
-        if (winner) wins[winner] = (wins[winner] || 0) + 1;
-    });
-    return wins;
-}
-
-function findPointsInconsistencies(leads, follows, rounds) {
-    const leadWins = computeRoundWinCounts(rounds, 'lead_winner');
-    const followWins = computeRoundWinCounts(rounds, 'follow_winner');
-    const issues = [];
-    const check = (list, wins, role) => {
-        (list || []).forEach((p) => {
-            if (!p || !p.name) return;
-            const recordedPoints = Number(p.points) || 0;
-            const roundWins = wins[p.name] || 0;
-            if (recordedPoints !== roundWins) {
-                issues.push({ role, name: p.name, points: recordedPoints, wins: roundWins });
-            }
-        });
-    };
-    check(leads, leadWins, 'Lead');
-    check(follows, followWins, 'Follow');
-    return issues;
-}
-
-function renderConsistencyWarnings(boxId, listId, issues) {
-    const box = document.getElementById(boxId);
-    const list = document.getElementById(listId);
-    if (!box || !list) return;
-    list.innerHTML = '';
-    if (!issues.length) {
-        box.classList.add('hidden');
-        return;
-    }
-    issues.forEach((issue) => {
-        const li = document.createElement('li');
-        li.textContent = `${issue.name} (${issue.role}): ${issue.points} point${issue.points === 1 ? '' : 's'} recorded, but ${issue.wins} round win${issue.wins === 1 ? '' : 's'} in the round history.`;
-        list.appendChild(li);
-    });
-    box.classList.remove('hidden');
-}
-
-// Replace every exact-match occurrence of `oldName` anywhere in the payload
-// tree (pairs, votes, tiebreak lists, champions, etc.) with `newName`, so a
-// contestant rename doesn't leave round history pointing at a stale name.
-function deepRenameContestant(node, oldName, newName) {
-    if (Array.isArray(node)) {
-        node.forEach((val, i) => {
-            if (val === oldName) node[i] = newName;
-            else if (val && typeof val === 'object') deepRenameContestant(val, oldName, newName);
-        });
-    } else if (node && typeof node === 'object') {
-        Object.keys(node).forEach((key) => {
-            const val = node[key];
-            if (val === oldName) node[key] = newName;
-            else if (val && typeof val === 'object') deepRenameContestant(val, oldName, newName);
-        });
-    }
-}
-
-// Mirrors the backend's placement ranking (_export_participants): sorted by
-// points desc, ties share a rank.
-function computePlacements(entries) {
-    const sorted = [...entries].sort((a, b) => (b.points || 0) - (a.points || 0));
-    let lastPoints = null;
-    let lastRank = 0;
-    sorted.forEach((c, idx) => {
-        if (c.points !== lastPoints) {
-            lastRank = idx + 1;
-            lastPoints = c.points;
-        }
-        c.placement = lastRank;
-    });
-}
-
-// Generalized battle-payload editor. `payload` is a raw hustlentussle.battle v1 object;
-// `options.onSave(editedPayload, meta)` decides what "save" means for the caller (e.g. the
-// results screen reprocesses+redisplays locally, while the YTD admin screen PUTs to the
-// server) - see editorCallbacks usage in saveEditedResults(). `options.showMetaFields` shows
-// battle name/date inputs (only meaningful for already-published battles, which have a
-// separate DB-level name/date from the payload itself); `options.initialMeta` pre-fills them.
-function openEditResultsModal(payload, options = {}) {
-    if (!payload) return;
-    editorPayload = payload;
-    editorCallbacks = options;
-
-    payload.participants = payload.participants || { leads: [], follows: [] };
-    payload.participants.leads = payload.participants.leads || [];
-    payload.participants.follows = payload.participants.follows || [];
-
-    const titleEl = document.getElementById('edit-results-title');
-    if (titleEl) titleEl.textContent = options.title || 'Edit Results';
-
-    const metaSection = document.getElementById('edit-results-meta');
-    if (metaSection) {
-        metaSection.style.display = options.showMetaFields ? '' : 'none';
-        if (options.showMetaFields) {
-            const nameInput = document.getElementById('edit-battle-name');
-            const dateInput = document.getElementById('edit-battle-date');
-            if (nameInput) nameInput.value = (options.initialMeta && options.initialMeta.name) || '';
-            if (dateInput) dateInput.value = (options.initialMeta && options.initialMeta.battle_date) || '';
-        }
-    }
-
-    renderEditParticipants('edit-leads-body', payload.participants.leads);
-    renderEditParticipants('edit-follows-body', payload.participants.follows);
-    renderEditChampions(payload);
-    renderEditRounds(payload);
-    refreshEditConsistencyWarnings();
-
-    const errorEl = document.getElementById('edit-results-error');
-    if (errorEl) { errorEl.textContent = ''; errorEl.classList.add('hidden'); }
-
-    const modal = document.getElementById('edit-results-modal');
-    if (modal) modal.classList.remove('hidden');
-}
-// Exposed so other modules (e.g. web/js/ytd.js, which is intentionally self-contained and
-// doesn't share scope with this file) can reuse this editor without duplicating its ~250
-// lines of table/select-building logic.
-window.openBattlePayloadEditor = openEditResultsModal;
-
-// Recomputes the points-vs-round-wins warning from whatever is currently typed/selected
-// in the modal (not the saved payload), so it updates live as the user edits.
-function refreshEditConsistencyWarnings() {
-    const leadEntries = collectParticipantEdits('edit-leads-body').entries;
-    const followEntries = collectParticipantEdits('edit-follows-body').entries;
-    const rounds = Array.from(document.querySelectorAll('#edit-rounds-body .edit-round-row')).map((rowEl) => {
-        const leadSel = rowEl.querySelector('.edit-round-lead-winner');
-        const followSel = rowEl.querySelector('.edit-round-follow-winner');
-        return {
-            lead_winner: leadSel ? (leadSel.value || null) : null,
-            follow_winner: followSel ? (followSel.value || null) : null,
-        };
-    });
-    const issues = findPointsInconsistencies(leadEntries, followEntries, rounds);
-    renderConsistencyWarnings('edit-consistency-warning', 'edit-consistency-list', issues);
-}
-
-function closeEditResultsModal() {
-    const modal = document.getElementById('edit-results-modal');
-    if (modal) modal.classList.add('hidden');
-}
-
-function buildEditParticipantRow(participant) {
-    const row = document.createElement('tr');
-    row.dataset.originalName = participant.name || '';
-
-    const nameCell = document.createElement('td');
-    const nameInput = document.createElement('input');
-    nameInput.type = 'text';
-    nameInput.className = 'edit-name-input';
-    nameInput.value = participant.name || '';
-    nameCell.appendChild(nameInput);
-
-    const pointsCell = document.createElement('td');
-    const pointsInput = document.createElement('input');
-    pointsInput.type = 'number';
-    pointsInput.min = '0';
-    pointsInput.step = '1';
-    pointsInput.className = 'edit-points-input';
-    pointsInput.value = participant.points ?? 0;
-    pointsCell.appendChild(pointsInput);
-
-    const removeCell = document.createElement('td');
-    const removeBtn = document.createElement('button');
-    removeBtn.type = 'button';
-    removeBtn.className = 'btn secondary small';
-    removeBtn.textContent = 'Remove';
-    removeBtn.addEventListener('click', () => { row.remove(); refreshEditConsistencyWarnings(); });
-    removeCell.appendChild(removeBtn);
-
-    row.appendChild(nameCell);
-    row.appendChild(pointsCell);
-    row.appendChild(removeCell);
-    return row;
-}
-
-function renderEditParticipants(tbodyId, participants) {
-    const tbody = document.getElementById(tbodyId);
-    if (!tbody) return;
-    tbody.innerHTML = '';
-    participants.forEach((p) => tbody.appendChild(buildEditParticipantRow(p)));
-}
-
-function fillChampionSelect(select, names, currentName) {
-    if (!select) return;
-    select.innerHTML = '';
-    select.appendChild(new Option('No champion', ''));
-    names.forEach((name) => select.appendChild(new Option(name, name, false, name === currentName)));
-    if (!currentName) select.value = '';
-}
-
-function renderEditChampions(payload) {
-    const leadNames = payload.participants.leads.map((p) => p.name).filter(Boolean);
-    const followNames = payload.participants.follows.map((p) => p.name).filter(Boolean);
-    fillChampionSelect(
-        document.getElementById('edit-champion-lead'), leadNames,
-        payload.champions && payload.champions.lead ? payload.champions.lead.name : null
-    );
-    fillChampionSelect(
-        document.getElementById('edit-champion-follow'), followNames,
-        payload.champions && payload.champions.follow ? payload.champions.follow.name : null
-    );
-}
-
-function buildEditRoundRow(round) {
-    const row = document.createElement('div');
-    row.className = 'edit-round-row';
-    row.dataset.roundNum = round.round_num;
-
-    const label = document.createElement('span');
-    label.className = 'edit-round-label';
-    label.textContent = round.tiebreak ? `Round ${round.round_num} (Tie-Break)` : `Round ${round.round_num}`;
-    row.appendChild(label);
-
-    const leadOptions = round.tiebreak
-        ? (round.tiebreak_leads || [])
-        : [round.pairs?.pair_1?.lead, round.pairs?.pair_2?.lead].filter(Boolean);
-    const followOptions = round.tiebreak
-        ? (round.tiebreak_follows || [])
-        : [round.pairs?.pair_1?.follow, round.pairs?.pair_2?.follow].filter(Boolean);
-
-    if (leadOptions.length) {
-        const leadSel = document.createElement('select');
-        leadSel.className = 'edit-round-lead-winner';
-        leadSel.appendChild(new Option('No lead winner', ''));
-        leadOptions.forEach((name) => leadSel.appendChild(new Option(name, name, false, name === round.lead_winner)));
-        row.appendChild(leadSel);
-    }
-    if (followOptions.length) {
-        const followSel = document.createElement('select');
-        followSel.className = 'edit-round-follow-winner';
-        followSel.appendChild(new Option('No follow winner', ''));
-        followOptions.forEach((name) => followSel.appendChild(new Option(name, name, false, name === round.follow_winner)));
-        row.appendChild(followSel);
-    }
-    return row;
-}
-
-function renderEditRounds(payload) {
-    const container = document.getElementById('edit-rounds-body');
-    if (!container) return;
-    container.innerHTML = '';
-    const rounds = (payload.rounds || []).slice().sort((a, b) => (a.round_num || 0) - (b.round_num || 0));
-    rounds.forEach((round) => container.appendChild(buildEditRoundRow(round)));
-}
-
-// Reads the editable rows for one role's table; blank names are treated as removed rows.
-function collectParticipantEdits(tbodyId) {
-    const tbody = document.getElementById(tbodyId);
-    const entries = [];
-    const renames = [];
-    Array.from(tbody.querySelectorAll('tr')).forEach((row) => {
-        const name = row.querySelector('.edit-name-input').value.trim();
-        if (!name) return;
-        const pointsRaw = parseInt(row.querySelector('.edit-points-input').value, 10);
-        const points = Number.isFinite(pointsRaw) ? pointsRaw : 0;
-        const originalName = row.dataset.originalName;
-        if (originalName && originalName !== name) renames.push({ from: originalName, to: name });
-        entries.push({ name, points });
-    });
-    return { entries, renames };
-}
-
-function validateParticipantEdits(entries, label) {
-    const seen = new Set();
-    entries.forEach((e) => {
-        if (seen.has(e.name)) throw new Error(`Duplicate ${label} name: "${e.name}"`);
-        seen.add(e.name);
-        if (e.points < 0) throw new Error(`${label} "${e.name}" cannot have negative points.`);
-    });
-}
-
-// Rebuilds payload.participants[role] from the edited rows, preserving
-// existing initial_order for known names and appending new ones at the end.
-// `championName` (already post-rename) drives is_champion, which is what
-// YTD stats ingest (stats/normalize.py) reads for the crown — the top-level
-// `champions` object alone only affects this screen's display.
-function applyParticipantEdits(payload, role, entries, championName) {
-    const existing = payload.participants[role] || [];
-    const byName = new Map(existing.map((p) => [p.name, p]));
-    const updated = entries.map((e, idx) => {
-        const match = byName.get(e.name);
-        const base = match || { name: e.name, placement: 0, initial_order: existing.length + idx + 1 };
-        base.points = e.points;
-        base.is_champion = championName != null && e.name === championName;
-        return base;
-    });
-    computePlacements(updated);
-    payload.participants[role] = updated;
-}
-
-// Re-derives the results-screen display shape by sending the edited payload
-// through the same endpoint/validation the original upload used.
-async function reprocessBattlePayload(payload) {
-    const blob = new Blob([JSON.stringify(payload)], { type: 'application/json' });
-    const formData = new FormData();
-    formData.append('battle_file', blob, 'edited-battle.json');
-    const response = await fetch('/api/process_uploaded_file', { method: 'POST', body: formData });
-    const data = await response.json();
-    if (!response.ok || data.error) {
-        throw new Error(data.error || `Failed to process edited results (${response.status})`);
-    }
-    return data;
-}
-
-async function saveEditedResults() {
-    const errorEl = document.getElementById('edit-results-error');
-    if (errorEl) { errorEl.textContent = ''; errorEl.classList.add('hidden'); }
-
-    const payload = editorPayload;
-    if (!payload || !editorCallbacks) return;
-
-    try {
-        const leadEdits = collectParticipantEdits('edit-leads-body');
-        const followEdits = collectParticipantEdits('edit-follows-body');
-        validateParticipantEdits(leadEdits.entries, 'Lead');
-        validateParticipantEdits(followEdits.entries, 'Follow');
-
-        // Apply the champion selection and round-winner edits using the *current*
-        // names first, then sweep renames afterwards so a freshly-picked name
-        // still gets corrected if that same row is also being renamed.
-        payload.champions = payload.champions || {};
-        const leadChampionName = document.getElementById('edit-champion-lead').value || null;
-        const followChampionName = document.getElementById('edit-champion-follow').value || null;
-        payload.champions.lead = leadChampionName
-            ? { ...(payload.champions.lead || {}), name: leadChampionName }
-            : null;
-        payload.champions.follow = followChampionName
-            ? { ...(payload.champions.follow || {}), name: followChampionName }
-            : null;
-
-        document.querySelectorAll('#edit-rounds-body .edit-round-row').forEach((rowEl) => {
-            const roundNum = Number(rowEl.dataset.roundNum);
-            const round = (payload.rounds || []).find((r) => r.round_num === roundNum);
-            if (!round) return;
-            const leadSel = rowEl.querySelector('.edit-round-lead-winner');
-            const followSel = rowEl.querySelector('.edit-round-follow-winner');
-            if (leadSel) round.lead_winner = leadSel.value || null;
-            if (followSel) round.follow_winner = followSel.value || null;
-        });
-
-        [...leadEdits.renames, ...followEdits.renames].forEach(({ from, to }) => {
-            if (from && to && from !== to) deepRenameContestant(payload, from, to);
-        });
-
-        // Re-read the champion names post-rename so a renamed champion still matches.
-        const finalLeadChampion = payload.champions.lead ? payload.champions.lead.name : null;
-        const finalFollowChampion = payload.champions.follow ? payload.champions.follow.name : null;
-        applyParticipantEdits(payload, 'leads', leadEdits.entries, finalLeadChampion);
-        applyParticipantEdits(payload, 'follows', followEdits.entries, finalFollowChampion);
-
-        const meta = editorCallbacks.showMetaFields
-            ? {
-                name: document.getElementById('edit-battle-name').value.trim(),
-                battle_date: document.getElementById('edit-battle-date').value,
-            }
-            : null;
-        await editorCallbacks.onSave(payload, meta);
-        closeEditResultsModal();
-    } catch (err) {
-        if (errorEl) {
-            errorEl.textContent = err.message || 'Failed to save changes.';
-            errorEl.classList.remove('hidden');
-        }
-    }
-}
-
 // --- Router hydration helpers (called by js/router.js) -------------------
-
-// Navigate to the results screen for a given payload, updating the URL.
-function showResults(data, opts) {
-    currentResultsData = data;
-    let sid;
-    if (opts && Object.prototype.hasOwnProperty.call(opts, 'sessionId')) sid = opts.sessionId;
-    else sid = (data && data.session_id) ? data.session_id : sessionId;
-    navigate(sid ? '/results/' + encodeURIComponent(sid) : '/results');
-}
 
 // Enter /battle/<id>: rebuild the interactive (or display) battle from server state.
 function hydrateBattleRoute(sid) {
@@ -2712,886 +2270,6 @@ function hydrateBattleRoute(sid) {
             try { showToast('That battle was not found or has expired.', 'error'); } catch (e) {}
             navigate('/', { replace: true });
         });
-}
-
-// Enter /results/<id> (or /results with in-memory data): render final results.
-function hydrateResultsRoute(sid) {
-    if (!sid) {
-        if (currentResultsData) displayResults(currentResultsData);
-        else navigate('/', { replace: true });
-        return;
-    }
-    if (currentResultsData && currentResultsData.session_id === sid) {
-        displayResults(currentResultsData);
-        return;
-    }
-    fetch(`/api/results?session_id=${encodeURIComponent(sid)}`)
-        .then(r => { if (!r.ok) throw new Error('not found'); return r.json(); })
-        .then(data => { currentResultsData = data; displayResults(data); })
-        .catch(() => {
-            try { showToast('Those results were not found or have expired.', 'error'); } catch (e) {}
-            navigate('/', { replace: true });
-        });
-}
-
-async function displayResults(data) {
-    console.log('Displaying results with data:', data);
-    
-    // Ensure display mode polling is stopped when showing results
-    stopDisplayPolling();
-    
-    // Use the showScreen function with the correct screen element
-    showScreen(resultsScreen);
-
-    // Editing is only offered for battles loaded via file upload (no live session behind them)
-    const editResultsBtn = document.getElementById('edit-results-btn');
-    if (editResultsBtn) {
-        editResultsBtn.style.display = (data.uploaded && currentUploadedBattlePayload) ? '' : 'none';
-    }
-
-    // Flag points that don't match the round history — uploaded/edited files can drift;
-    // a live battle's points always come straight from the game engine, so skip the noise there.
-    const resultsIssues = data.uploaded ? findPointsInconsistencies(data.leads, data.follows, data.rounds) : [];
-    renderConsistencyWarnings('results-consistency-warning', 'results-consistency-list', resultsIssues);
-
-    // Clear previous results
-    const leadResultsBody = document.getElementById('lead-results-body');
-    const followResultsBody = document.getElementById('follow-results-body');
-    const roundsContainer = document.getElementById('rounds-accordion');
-    const leadsInitialOrder = document.getElementById('leads-initial-order');
-    const followsInitialOrder = document.getElementById('follows-initial-order');
-    const leadGraphic = document.getElementById('lead-graphic');
-    const followGraphic = document.getElementById('follow-graphic');
-    
-    if (!leadResultsBody || !followResultsBody || !roundsContainer) {
-        console.error('Could not find required elements for results display');
-        return;
-    }
-    
-    // Clear all previous data
-    leadResultsBody.innerHTML = '';
-    followResultsBody.innerHTML = '';
-    roundsContainer.innerHTML = '';
-    if (leadsInitialOrder) leadsInitialOrder.innerHTML = '';
-    if (followsInitialOrder) followsInitialOrder.innerHTML = '';
-    if (leadGraphic) leadGraphic.innerHTML = '';
-    if (followGraphic) followGraphic.innerHTML = '';
-    
-    // Sync globals so exportSocialImage() has current data
-    currentLeads = data.leads || [];
-    currentFollows = data.follows || [];
-
-    // Always use the initial order from the server response
-    const initialLeadsData = data.initial_leads || [];
-    const initialFollowsData = data.initial_follows || [];
-    console.log('Using initial leads from server:', initialLeadsData);
-    console.log('Using initial follows from server:', initialFollowsData);
-    
-    console.log('Lead results:', data.leads);
-    console.log('Follow results:', data.follows);
-    
-    // Crown the resolved battle champion (threshold / tie-break / early-end leader).
-    // Falls back to the top scorer only if champion info is absent.
-    const champions = data.champions;
-    function crownName(role, list) {
-        if (champions && champions[role] !== undefined) {
-            // Champion info present: trust it — a null name means no outright winner.
-            return champions[role] && champions[role].name ? champions[role].name : null;
-        }
-        // Legacy payload without champions: fall back to the top scorer.
-        if (Array.isArray(list) && list.length > 0) {
-            const top = list.reduce((best, c) =>
-                (Number(c.points) || 0) > (Number(best.points) || 0) ? c : best, list[0]);
-            return top && top.name ? top.name : null;
-        }
-        return null;
-    }
-    const topLeadName = crownName('lead', data.leads);
-    const topFollowName = crownName('follow', data.follows);
-
-    // Display lead results
-    if (data.leads && Array.isArray(data.leads)) {
-        data.leads.forEach(lead => {
-            const row = document.createElement('tr');
-            const nameCell = document.createElement('td');
-            const leadCrown = (topLeadName && lead.name === topLeadName) ? ' 👑' : '';
-            nameCell.textContent = `${lead.name}${leadCrown}`.trim();
-            
-            const pointsCell = document.createElement('td');
-            pointsCell.textContent = lead.points;
-            
-            row.appendChild(nameCell);
-            row.appendChild(pointsCell);
-            leadResultsBody.appendChild(row);
-        });
-    } else {
-        console.warn('No lead results data available');
-    }
-    
-    // Display follow results
-    if (data.follows && Array.isArray(data.follows)) {
-        data.follows.forEach(follow => {
-            const row = document.createElement('tr');
-            const nameCell = document.createElement('td');
-            const followCrown = (topFollowName && follow.name === topFollowName) ? ' 👑' : '';
-            nameCell.textContent = `${follow.name}${followCrown}`.trim();
-            
-            const pointsCell = document.createElement('td');
-            pointsCell.textContent = follow.points;
-            
-            row.appendChild(nameCell);
-            row.appendChild(pointsCell);
-            followResultsBody.appendChild(row);
-        });
-    } else {
-        console.warn('No follow results data available');
-    }
-
-    // Build battle graphic data: map contestant to rounds and wins
-    if (Array.isArray(data.rounds) && data.rounds.length > 0 && leadGraphic && followGraphic) {
-        const leadMap = new Map();
-        const followMap = new Map();
-        const totalRounds = data.rounds.length;
-        data.rounds.forEach(r => {
-            const roundNum = r.round_num;
-            if (r.tiebreak) {
-                (r.tiebreak_leads || []).forEach(name => {
-                    if (!leadMap.has(name)) leadMap.set(name, []);
-                    leadMap.get(name).push({ round: roundNum, win: r.lead_winner === name });
-                });
-                (r.tiebreak_follows || []).forEach(name => {
-                    if (!followMap.has(name)) followMap.set(name, []);
-                    followMap.get(name).push({ round: roundNum, win: r.follow_winner === name });
-                });
-                return;
-            }
-            const pair1Lead = r.pairs?.pair_1?.lead;
-            const pair1Follow = r.pairs?.pair_1?.follow;
-            const pair2Lead = r.pairs?.pair_2?.lead;
-            const pair2Follow = r.pairs?.pair_2?.follow;
-            const leadWinner = r.lead_winner;
-            const followWinner = r.follow_winner;
-            if (pair1Lead) {
-                if (!leadMap.has(pair1Lead)) leadMap.set(pair1Lead, []);
-                leadMap.get(pair1Lead).push({ round: roundNum, win: leadWinner === pair1Lead });
-            }
-            if (pair2Lead) {
-                if (!leadMap.has(pair2Lead)) leadMap.set(pair2Lead, []);
-                leadMap.get(pair2Lead).push({ round: roundNum, win: leadWinner === pair2Lead });
-            }
-            if (pair1Follow) {
-                if (!followMap.has(pair1Follow)) followMap.set(pair1Follow, []);
-                followMap.get(pair1Follow).push({ round: roundNum, win: followWinner === pair1Follow });
-            }
-            if (pair2Follow) {
-                if (!followMap.has(pair2Follow)) followMap.set(pair2Follow, []);
-                followMap.get(pair2Follow).push({ round: roundNum, win: followWinner === pair2Follow });
-            }
-        });
-
-        // Helper to render a column by initial order
-        const renderGraphicColumn = (initialOrder, dataMap, topName, container) => {
-            initialOrder.forEach((name, idx) => {
-                const row = document.createElement('div');
-                row.className = 'graphic-row';
-                const rank = document.createElement('div');
-                rank.className = 'graphic-rank';
-                rank.textContent = `${idx + 1}.`;
-                const nameDiv = document.createElement('div');
-                nameDiv.className = 'graphic-name';
-                nameDiv.textContent = name;
-                if (topName && name === topName) {
-                    const crown = document.createElement('span');
-                    crown.className = 'crown-icon';
-                    crown.textContent = '👑';
-                    nameDiv.appendChild(crown);
-                }
-                const badges = document.createElement('div');
-                badges.className = 'round-badges';
-                const rounds = dataMap.get(name) || [];
-                // Sort rounds ascending by round number
-                rounds.sort((a, b) => a.round - b.round);
-                rounds.forEach(info => {
-                    const b = document.createElement('div');
-                    b.className = 'badge' + (info.win ? ' win' : '');
-                    b.textContent = info.round;
-                    badges.appendChild(b);
-                });
-                row.appendChild(rank);
-                row.appendChild(nameDiv);
-                row.appendChild(badges);
-                container.appendChild(row);
-            });
-        };
-
-        renderGraphicColumn(initialLeadsData || [], leadMap, topLeadName, leadGraphic);
-        renderGraphicColumn(initialFollowsData || [], followMap, topFollowName, followGraphic);
-
-        // Cache for social export
-        resultsLeadMap = leadMap;
-        resultsFollowMap = followMap;
-        resultsInitialLeads = [...(initialLeadsData || [])];
-        resultsInitialFollows = [...(initialFollowsData || [])];
-        resultsTopLeadName = topLeadName;
-        resultsTopFollowName = topFollowName;
-        resultsNumRounds = totalRounds;
-        resultsGuestJudges = guestJudges.length > 0
-            ? [...guestJudges]
-            : (() => {
-                // Fallback: extract unique judge names from rounds data
-                // (handles page-reload case where guestJudges global was reset)
-                const seen = new Set();
-                (data.rounds || []).forEach(r => {
-                    if (Array.isArray(r.judges)) r.judges.forEach(j => seen.add(j));
-                });
-                return [...seen];
-            })();
-    }
-    
-    console.log('Round history:', data.rounds);
-    
-    // Always render round history; optionally enrich with Spotify metadata if enabled
-    const spotifyOn = localStorage.getItem('spotify.enabled') === 'true';
-    if (spotifyOn && data.rounds && data.rounds.length > 0) {
-        try {
-            const access_token = await getSpotifyToken();
-            
-            // Fetch metadata for all rounds in parallel
-            await Promise.all(data.rounds.map(async (round) => {
-                if (spotifyOn && round.song_info && round.song_info.spotify_url) {
-                    try {
-                        const spotifyUrl = new URL(round.song_info.spotify_url);
-                        const trackId = spotifyUrl.pathname.split('/').pop();
-                        
-                        if (trackId) {
-                            const metadataResponse = await fetch(`https://api.spotify.com/v1/tracks/${trackId}`, {
-                                headers: {
-                                    'Authorization': `Bearer ${access_token}`
-                                }
-                            });
-                            
-                            if (metadataResponse.status === 401 && spotifyOn) {
-                                // Token expired, get a new one and retry
-                                const newToken = await getSpotifyToken();
-                                const retryResponse = await fetch(`https://api.spotify.com/v1/tracks/${trackId}`, {
-                                    headers: {
-                                        'Authorization': `Bearer ${newToken}`
-                                    }
-                                });
-                                
-                                if (retryResponse.ok) {
-                                    const metadata = await retryResponse.json();
-                                    round.song_info.title = metadata.name;
-                                    round.song_info.artist = metadata.artists.map(artist => artist.name).join(', ');
-                                }
-                            } else if (metadataResponse.ok) {
-                                const metadata = await metadataResponse.json();
-                                round.song_info.title = metadata.name;
-                                round.song_info.artist = metadata.artists.map(artist => artist.name).join(', ');
-                            }
-                        }
-                    } catch (e) {
-                        console.error(`Error fetching metadata for ${round.song_info.spotify_url}:`, e);
-                    }
-                }
-            }));
-        } catch (error) {
-            console.error('Error fetching Spotify metadata:', error);
-        }
-        
-        // Fallthrough to render below
-    }
-
-    // Display round history regardless of Spotify status
-    if (data.rounds && data.rounds.length > 0) {
-        displayRoundHistory(data.rounds);
-    } else {
-        console.warn('No round history data available');
-    }
-
-    showToast('Battle complete!', 'success');
-}
-
-// Update the displayRoundHistory function
-function displayRoundHistory(rounds) {
-    const roundsContainer = document.getElementById('rounds-accordion');
-    roundsContainer.innerHTML = '';
-    
-    // Sort rounds by round number
-    rounds.sort((a, b) => a.round_num - b.round_num);
-    
-    // Create an accordion item for each round
-    rounds.forEach(round => {
-        const accordionItem = document.createElement('div');
-        accordionItem.className = 'accordion-item' + (round.tiebreak ? ' tiebreak' : '');
-
-        // Create the header
-        const header = document.createElement('div');
-        header.className = 'accordion-header' + (round.tiebreak ? ' tiebreak' : '');
-        const headerLabel = round.tiebreak
-            ? `Round ${round.round_num} <span class="tiebreak-history-badge">Tie-Break</span>`
-            : `Round ${round.round_num}`;
-        header.innerHTML = `<span>${headerLabel}</span><span>+</span>`;
-        
-        // Add click handler for accordion functionality
-        header.addEventListener('click', () => {
-            const content = header.nextElementSibling;
-            const isOpen = content.classList.contains('open');
-
-            // Toggle open state
-            header.classList.toggle('active');
-            content.classList.toggle('open');
-
-            // Update the plus/minus symbol
-            const symbol = header.lastElementChild;
-            symbol.textContent = isOpen ? '+' : '-';
-        });
-        
-        // Create the content
-        const content = document.createElement('div');
-        content.className = 'accordion-content';
-        
-        // Add round details
-        const details = document.createElement('div');
-        details.className = 'round-details';
-        const topRow = document.createElement('div');
-        topRow.className = 'round-top-row';
-        
-        // Add song information if available
-        if (round.song_info) {
-            const songSection = document.createElement('div');
-            songSection.className = 'round-song';
-            let songHTML = '<h4>Song</h4>';
-            
-            // Add song title and artist if available
-            if (round.song_info.title || round.song_info.artist) {
-                songHTML += '<div class="song-details">';
-                if (round.song_info.title) {
-                    songHTML += `<div class="song-title">${round.song_info.title}</div>`;
-                }
-                if (round.song_info.artist) {
-                    songHTML += `<div class="song-artist">${round.song_info.artist}</div>`;
-                }
-                songHTML += '</div>';
-            }
-            
-            if (localStorage.getItem('spotify.enabled') === 'true' && round.song_info.spotify_url) {
-                try {
-                    const spotifyUrl = new URL(round.song_info.spotify_url);
-                    const trackId = spotifyUrl.pathname.split('/').pop();
-                    if (trackId) {
-                        songHTML += `
-                            <iframe 
-                                src="https://open.spotify.com/embed/track/${trackId}" 
-                                width="100%" 
-                                height="80" 
-                                frameborder="0" 
-                                allowtransparency="true" 
-                                allow="encrypted-media"
-                                class="spotify-embed">
-                            </iframe>`;
-                    }
-                } catch (e) {
-                    console.error('Invalid Spotify URL:', e);
-                }
-            }
-            
-            songSection.innerHTML = songHTML;
-            topRow.appendChild(songSection);
-        }
-        
-        // Participants section
-        const participants = document.createElement('div');
-        participants.className = 'round-participants';
-
-        if (round.tiebreak) {
-            let html = '<h4>Tie-Break</h4>';
-            if (round.tiebreak_leads?.length) {
-                html += `<div class="tiebreak-history-group">
-                    <span class="tiebreak-history-role-label">Leads</span>
-                    <span class="tiebreak-history-names">
-                        ${round.tiebreak_leads.map(n =>
-                            `<span class="contestant lead${round.lead_winner === n ? ' tiebreak-history-winner' : ''}">${n}${round.lead_winner === n ? ' 🏆' : ''}</span>`
-                        ).join('<span class="tiebreak-history-vs">vs</span>')}
-                    </span>
-                </div>`;
-            }
-            if (round.tiebreak_follows?.length) {
-                html += `<div class="tiebreak-history-group">
-                    <span class="tiebreak-history-role-label">Follows</span>
-                    <span class="tiebreak-history-names">
-                        ${round.tiebreak_follows.map(n =>
-                            `<span class="contestant follow${round.follow_winner === n ? ' tiebreak-history-winner' : ''}">${n}${round.follow_winner === n ? ' 🏆' : ''}</span>`
-                        ).join('<span class="tiebreak-history-vs">vs</span>')}
-                    </span>
-                </div>`;
-            }
-            participants.innerHTML = html;
-            topRow.appendChild(participants);
-            details.appendChild(topRow);
-        } else {
-            let participantsHTML = '<h4>Participants</h4>';
-            if (round.pairs && Object.keys(round.pairs).length > 0) {
-                participantsHTML += `
-                    <div class="match-pair">
-                        <div>Couple 1:</div>
-                        <div><span class="lead">${round.pairs.pair_1.lead}</span> (Lead) &
-                        <span class="follow">${round.pairs.pair_1.follow}</span> (Follow)</div>
-                    </div>
-                    <div class="match-pair">
-                        <div>Couple 2:</div>
-                        <div><span class="lead">${round.pairs.pair_2.lead}</span> (Lead) &
-                        <span class="follow">${round.pairs.pair_2.follow}</span> (Follow)</div>
-                    </div>
-                `;
-                if (round.lead_winner) participantsHTML += `<div class="winner">Lead Winner: ${round.lead_winner}</div>`;
-                if (round.follow_winner) participantsHTML += `<div class="winner">Follow Winner: ${round.follow_winner}</div>`;
-            } else {
-                participantsHTML += '<p>No participant data available for this round.</p>';
-            }
-            participants.innerHTML = participantsHTML;
-            topRow.appendChild(participants);
-            details.appendChild(topRow);
-        }
-
-        // Add judge votes section (normal rounds only)
-        if (round.tiebreak) {
-            content.appendChild(details);
-            accordionItem.appendChild(header);
-            accordionItem.appendChild(content);
-            roundsContainer.appendChild(accordionItem);
-            return;
-        }
-
-        const judgeVotes = document.createElement('div');
-        judgeVotes.className = 'judge-votes';
-        let judgeVotesHTML = '<h4>Judge Votes</h4>';
-        judgeVotesHTML += '<div class="vote-sections">';
-
-        // Lead votes
-        if (round.lead_votes) {
-            judgeVotesHTML += '<div class="vote-section lead"><h5>Lead Votes</h5>';
-            
-            // Sort votes to show guest judges first
-            const sortedVotes = Object.entries(round.lead_votes).sort((a, b) => {
-                const aIsGuest = guestJudges.includes(a[0]);
-                const bIsGuest = guestJudges.includes(b[0]);
-                if (aIsGuest && !bIsGuest) return -1;
-                if (!aIsGuest && bIsGuest) return 1;
-                return 0;
-            });
-
-            judgeVotesHTML += '<table class="judge-votes-table"><thead><tr><th>Judge</th><th>Type</th><th>Vote</th></tr></thead><tbody>';
-            sortedVotes.forEach(([judge, vote]) => {
-                const isGuest = guestJudges.includes(judge);
-                const voteText = getVoteText(vote, round);
-                const judgeType = isGuest ? 'Guest' : 'Contestant';
-                judgeVotesHTML += `
-                    <tr class="judge-vote ${isGuest ? 'guest-judge' : ''}">
-                        <td class="judge-name">${judge}</td>
-                        <td><span class="judge-type ${isGuest ? 'guest' : 'contestant'}">${judgeType}</span></td>
-                        <td class="vote">${voteText}</td>
-                    </tr>
-                `;
-            });
-            judgeVotesHTML += '</tbody></table></div>';
-        }
-
-        // Follow votes
-        if (round.follow_votes) {
-            judgeVotesHTML += '<div class="vote-section follow"><h5>Follow Votes</h5>';
-            
-            // Sort votes to show guest judges first
-            const sortedVotes = Object.entries(round.follow_votes).sort((a, b) => {
-                const aIsGuest = guestJudges.includes(a[0]);
-                const bIsGuest = guestJudges.includes(b[0]);
-                if (aIsGuest && !bIsGuest) return -1;
-                if (!aIsGuest && bIsGuest) return 1;
-                return 0;
-            });
-
-            judgeVotesHTML += '<table class="judge-votes-table"><thead><tr><th>Judge</th><th>Type</th><th>Vote</th></tr></thead><tbody>';
-            sortedVotes.forEach(([judge, vote]) => {
-                const isGuest = guestJudges.includes(judge);
-                const voteText = getFollowVoteText(vote, round);
-                const judgeType = isGuest ? 'Guest' : 'Contestant';
-                judgeVotesHTML += `
-                    <tr class="judge-vote ${isGuest ? 'guest-judge' : ''}">
-                        <td class="judge-name">${judge}</td>
-                        <td><span class="judge-type ${isGuest ? 'guest' : 'contestant'}">${judgeType}</span></td>
-                        <td class="vote">${voteText}</td>
-                    </tr>
-                `;
-            });
-            judgeVotesHTML += '</tbody></table></div>';
-        }
-        judgeVotesHTML += '</div>';
-        judgeVotes.innerHTML = judgeVotesHTML;
-        details.appendChild(judgeVotes);
-        
-        // Session ID removed from round cell for cleaner display
-        
-        // Append assembled sections
-        content.appendChild(details);
-        accordionItem.appendChild(header);
-        accordionItem.appendChild(content);
-        roundsContainer.appendChild(accordionItem);
-    });
-}
-
-// Helper function to convert vote numbers to text
-function getVoteText(vote, round) {
-    if (!round || !round.pairs) return 'Unknown';
-    
-    switch (vote) {
-        case 1:
-            return round.pairs.pair_1.lead;
-        case 2:
-            return round.pairs.pair_2.lead;
-        case 3:
-            return 'Tie';
-        case 4:
-            return 'No Contest';
-        default:
-            return 'Unknown';
-    }
-}
-
-// Helper function to convert follow vote numbers to text
-function getFollowVoteText(vote, round) {
-    if (!round || !round.pairs) return 'Unknown';
-    
-    switch (vote) {
-        case 1:
-            return round.pairs.pair_1.follow;
-        case 2:
-            return round.pairs.pair_2.follow;
-        case 3:
-            return 'Tie';
-        case 4:
-            return 'No Contest';
-        default:
-            return 'Unknown';
-    }
-}
-
-function _rrect(ctx, x, y, w, h, r) {
-    ctx.beginPath();
-    ctx.moveTo(x + r, y);
-    ctx.lineTo(x + w - r, y);
-    ctx.quadraticCurveTo(x + w, y, x + w, y + r);
-    ctx.lineTo(x + w, y + h - r);
-    ctx.quadraticCurveTo(x + w, y + h, x + w - r, y + h);
-    ctx.lineTo(x + r, y + h);
-    ctx.quadraticCurveTo(x, y + h, x, y + h - r);
-    ctx.lineTo(x, y + r);
-    ctx.quadraticCurveTo(x, y, x + r, y);
-    ctx.closePath();
-}
-
-async function exportSocialImage() {
-    const W = 1080;
-    const H = 1920;
-    const PAD = 50;
-
-    await document.fonts.ready;
-
-    // Mirror the app's current light/dark theme
-    const isDark = document.documentElement.getAttribute('data-color') === 'dark';
-    const C = isDark ? {
-        bg:            '#0a0a12',
-        bgCard:        '#1a1a2e',
-        accent:        '#7c3aed',
-        textPrimary:   '#f1f5f9',
-        textSecondary: '#94a3b8',
-        textMuted:     '#64748b',
-        border:        'rgba(148,163,184,0.12)',
-        rowAlt:        'rgba(255,255,255,0.03)',
-        badgeWin:      '#7c3aed',
-        badgeWinBorder:'#9d5cf5',
-        badgeLose:     'rgba(255,255,255,0.07)',
-        badgeLoseBorder:'rgba(148,163,184,0.15)',
-        badgeWinText:  '#ffffff',
-        badgeLoseText: '#64748b',
-        fontDisplay:   '"Space Grotesk","Inter",sans-serif',
-        fontBody:      '"Inter","DM Sans",sans-serif',
-        fontMono:      '"DM Mono",monospace',
-    } : {
-        bg:            '#f5f5f7',
-        bgCard:        '#ffffff',
-        accent:        '#1d4ed8',
-        textPrimary:   '#0f172a',
-        textSecondary: '#475569',
-        textMuted:     '#94a3b8',
-        border:        '#e2e8f0',
-        rowAlt:        'rgba(0,0,0,0.03)',
-        badgeWin:      '#1d4ed8',
-        badgeWinBorder:'#1e40af',
-        badgeLose:     '#f0f0f5',
-        badgeLoseBorder:'#e2e8f0',
-        badgeWinText:  '#ffffff',
-        badgeLoseText: '#94a3b8',
-        fontDisplay:   '"DM Sans",sans-serif',
-        fontBody:      '"DM Sans",sans-serif',
-        fontMono:      '"DM Mono",monospace',
-    };
-
-    const canvas = document.createElement('canvas');
-    canvas.width = W;
-    canvas.height = H;
-    const ctx = canvas.getContext('2d');
-
-    // Background
-    ctx.fillStyle = C.bg;
-    ctx.fillRect(0, 0, W, H);
-
-    // Top accent bar
-    ctx.fillStyle = C.accent;
-    ctx.fillRect(0, 0, W, 12);
-
-    // --- Header ---
-    ctx.textAlign = 'center';
-    ctx.fillStyle = C.textPrimary;
-    ctx.font = `bold 78px ${C.fontDisplay}`;
-    ctx.fillText("Hustle n' Tussle", W / 2, 130);
-
-    const now = new Date();
-    const subtitleText = resultsEventTitle
-        || `${now.toLocaleDateString('en-US', { month: 'long' })} Edition (${now.getFullYear()})`;
-    ctx.fillStyle = C.textSecondary;
-    ctx.font = `400 38px ${C.fontBody}`;
-    ctx.fillText(subtitleText, W / 2, 200);
-
-    if (resultsGuestJudges.length > 0) {
-        const judgeLabel = resultsGuestJudges.length === 1 ? 'Judge' : 'Judges';
-        ctx.fillStyle = C.textSecondary;
-        ctx.font = `600 40px ${C.fontDisplay}`;
-        ctx.fillText(`${judgeLabel}: ${resultsGuestJudges.join(', ')}`, W / 2, 258);
-    }
-
-    const sortedLeads = [...currentLeads].sort((a, b) => (b.points || 0) - (a.points || 0));
-    const sortedFollows = [...currentFollows].sort((a, b) => (b.points || 0) - (a.points || 0));
-
-    const contentStartY = 300;
-
-    // --- Battle Graphic ---
-    const leadsOrder = resultsInitialLeads.length ? resultsInitialLeads : sortedLeads.map(l => l.name);
-    const followsOrder = resultsInitialFollows.length ? resultsInitialFollows : sortedFollows.map(f => f.name);
-    const numRounds = resultsNumRounds || 0;
-
-    // Layout: two card sections (Leads then Follows)
-    const CARD_HEADER_H = 76;   // dark strip at top of each card
-    const CARD_PAD_BOTTOM = 16; // padding below last row inside card
-    const SECTION_GAP = 20;
-    const FOOTER_H = 40;
-    const availForRows = (H - FOOTER_H) - contentStartY
-        - 2 * (CARD_HEADER_H + CARD_PAD_BOTTOM)
-        - SECTION_GAP;
-    const maxDancers = Math.max(leadsOrder.length, followsOrder.length);
-
-    // Badge sizing: fit all rounds on a single horizontal line per row
-    const rankW = 44;
-    const nameAreaW = 260;
-    const badgeStartX = PAD + rankW + nameAreaW + 24;
-    const badgeAreaW = W - PAD - badgeStartX;
-    const badgeGap = 4;
-
-    // Use the most rounds any individual dancer competed in (not total game rounds)
-    // so the badge row fills the full width for the most active dancer.
-    const allRoundCounts = [
-        ...[...leadsOrder].map(n => (resultsLeadMap.get(n) || []).length),
-        ...[...followsOrder].map(n => (resultsFollowMap.get(n) || []).length),
-    ];
-    const maxRoundsPerDancer = Math.max(...allRoundCounts, 1);
-    const badgeSizeFromRounds = Math.floor((badgeAreaW + badgeGap) / maxRoundsPerDancer) - badgeGap;
-
-    // rowH fills all available vertical space; badgeSize is the smaller of the two constraints
-    const maxRowHFromSpace = maxDancers > 0 ? Math.floor(availForRows / (2 * maxDancers)) : 70;
-    const rowH = Math.max(36, maxRowHFromSpace);
-    const badgeSize = Math.max(16, Math.min(rowH - 18, badgeSizeFromRounds));
-    const bFontSize = Math.max(9, Math.round(badgeSize * 0.52));
-    const nameFontSize = Math.max(20, Math.min(44, rowH - 26));
-    const rankFontSize = Math.max(16, nameFontSize - 4);
-
-    const drawSection = (order, map, topName, label, startY) => {
-        const CARD_RADIUS = 20;
-        const cardX = PAD - 12;
-        const cardW = W - 2 * (PAD - 12);
-        const cardH = CARD_HEADER_H + order.length * rowH + CARD_PAD_BOTTOM;
-
-        // Card background
-        _rrect(ctx, cardX, startY, cardW, cardH, CARD_RADIUS);
-        ctx.fillStyle = C.bgCard;
-        ctx.fill();
-
-        // Accent header strip — clipped to card shape so top corners are rounded
-        ctx.save();
-        _rrect(ctx, cardX, startY, cardW, cardH, CARD_RADIUS);
-        ctx.clip();
-        ctx.fillStyle = C.accent;
-        ctx.fillRect(cardX, startY, cardW, CARD_HEADER_H);
-        ctx.restore();
-
-        // Card border
-        _rrect(ctx, cardX, startY, cardW, cardH, CARD_RADIUS);
-        ctx.strokeStyle = C.border;
-        ctx.lineWidth = 1.5;
-        ctx.stroke();
-
-        // Section label — white bold uppercase in header strip
-        ctx.fillStyle = '#ffffff';
-        ctx.font = `bold 52px ${C.fontDisplay}`;
-        ctx.textAlign = 'left';
-        ctx.fillText(label.toUpperCase(), cardX + 24, startY + CARD_HEADER_H - 18);
-
-        const rowsStartY = startY + CARD_HEADER_H;
-
-        order.forEach((name, idx) => {
-            const rowY = rowsStartY + idx * rowH;
-            const isTop = name === topName;
-            const textBaseY = rowY + rowH * 0.64;
-
-            // Alternating row tint
-            if (idx % 2 === 0) {
-                ctx.fillStyle = C.rowAlt;
-                ctx.fillRect(PAD - 8, rowY + 1, W - (PAD - 8) * 2, rowH - 1);
-            }
-
-            // Rank
-            ctx.fillStyle = C.textMuted;
-            ctx.font = `400 ${rankFontSize}px ${C.fontMono}`;
-            ctx.textAlign = 'left';
-            ctx.fillText((idx + 1) + '.', PAD, textBaseY);
-
-            // Name + crown — clipped to nameAreaW so crown never bleeds into badge area
-            ctx.save();
-            ctx.beginPath();
-            ctx.rect(PAD + rankW, rowY, nameAreaW, rowH);
-            ctx.clip();
-            const maxChars = Math.floor(nameAreaW / (nameFontSize * 0.54));
-            const displayName = name.length > maxChars ? name.slice(0, maxChars - 1) + '…' : name;
-            ctx.fillStyle = isTop ? C.textPrimary : C.textSecondary;
-            ctx.font = isTop
-                ? `bold ${nameFontSize}px ${C.fontBody}`
-                : `400 ${nameFontSize}px ${C.fontBody}`;
-            ctx.fillText(displayName, PAD + rankW, textBaseY);
-            if (isTop) {
-                const nameW2 = ctx.measureText(displayName).width;
-                ctx.font = `${nameFontSize}px serif`;
-                ctx.fillText('👑', PAD + rankW + nameW2 + 5, textBaseY);
-            }
-            ctx.restore();
-
-            // Round badges — one row of circles
-            const rounds = (map.get(name) || []).slice().sort((a, b) => a.round - b.round);
-            const badgeCY = rowY + rowH / 2;
-            rounds.forEach((info, bi) => {
-                const cx = badgeStartX + bi * (badgeSize + badgeGap) + badgeSize / 2;
-                const cy = badgeCY;
-                const r = badgeSize / 2;
-
-                ctx.beginPath();
-                ctx.arc(cx, cy, r, 0, Math.PI * 2);
-                ctx.fillStyle = info.win ? C.badgeWin : C.badgeLose;
-                ctx.fill();
-
-                ctx.beginPath();
-                ctx.arc(cx, cy, r - 0.75, 0, Math.PI * 2);
-                ctx.strokeStyle = info.win ? C.badgeWinBorder : C.badgeLoseBorder;
-                ctx.lineWidth = 1.5;
-                ctx.stroke();
-
-                ctx.fillStyle = info.win ? C.badgeWinText : C.badgeLoseText;
-                ctx.font = `bold ${bFontSize}px ${C.fontMono}`;
-                ctx.textAlign = 'center';
-                ctx.fillText(String(info.round), cx, cy + bFontSize * 0.37);
-            });
-        });
-
-        return rowsStartY + order.length * rowH + CARD_PAD_BOTTOM;
-    };
-
-    const leadsEnd = drawSection(leadsOrder, resultsLeadMap, resultsTopLeadName, 'Leads', contentStartY);
-    drawSection(followsOrder, resultsFollowMap, resultsTopFollowName, 'Follows', leadsEnd + SECTION_GAP);
-
-    canvas.toBlob(blob => {
-        if (!blob) { showToast('Failed to generate image', 'error'); return; }
-        const url = URL.createObjectURL(blob);
-        const a = document.createElement('a');
-        a.href = url;
-        a.download = 'hnt-results.png';
-        document.body.appendChild(a);
-        a.click();
-        URL.revokeObjectURL(url);
-        document.body.removeChild(a);
-    }, 'image/png');
-}
-
-async function downloadBattleData() {
-    if (!sessionId && !currentUploadedBattlePayload) {
-        console.error('No active session or uploaded battle to download data from.');
-        return;
-    }
-
-    try {
-        // Live sessions fetch the portable JSON export; uploaded (possibly edited) battles
-        // already hold that same shape in memory.
-        let battleData;
-        if (sessionId) {
-            const response = await fetch(`/api/export_battle_data?session_id=${sessionId}`);
-            if (!response.ok) {
-                throw new Error(`Failed to fetch battle data: ${response.status}`);
-            }
-            battleData = await response.json();
-        } else {
-            battleData = currentUploadedBattlePayload;
-        }
-
-        // If Spotify integration is enabled, enrich song title/artist in-place
-        const spotifyOn = localStorage.getItem('spotify.enabled') === 'true';
-        if (spotifyOn) {
-            let access_token = await getSpotifyToken();
-            await Promise.all((battleData.rounds || []).map(async (round) => {
-                const song = round.song;
-                if (!song || !song.spotify_url) return;
-                try {
-                    const trackId = new URL(song.spotify_url).pathname.split('/').pop();
-                    if (!trackId) return;
-                    let md = await fetch(`https://api.spotify.com/v1/tracks/${trackId}`, {
-                        headers: { 'Authorization': `Bearer ${access_token}` }
-                    });
-                    if (md.status === 401) {
-                        access_token = await getSpotifyToken();
-                        md = await fetch(`https://api.spotify.com/v1/tracks/${trackId}`, {
-                            headers: { 'Authorization': `Bearer ${access_token}` }
-                        });
-                    }
-                    if (md.ok) {
-                        const meta = await md.json();
-                        song.title = meta.name;
-                        song.artist = meta.artists.map(a => a.name).join(', ');
-                    }
-                } catch (e) {
-                    console.error(`Error fetching metadata for ${song.spotify_url}:`, e);
-                }
-            }));
-        }
-
-        // Download the JSON directly
-        const blob = new Blob([JSON.stringify(battleData, null, 2)], { type: 'application/json' });
-        const url = window.URL.createObjectURL(blob);
-        const a = document.createElement('a');
-        a.href = url;
-        a.download = sessionId ? `battle_${sessionId}.json` : 'battle_edited.json';
-        document.body.appendChild(a);
-        a.click();
-        window.URL.revokeObjectURL(url);
-        document.body.removeChild(a);
-
-    } catch (error) {
-        console.error('Error downloading battle data:', error);
-        showToast('Failed to download battle data', 'error');
-    }
 }
 
 // End game function
